move css into component subfolders
